Maximize polymorphism and code reuse with hierarchical exceptions
**Task**
Define the `Shell` class with the following information:
- The file system it will work with, represented by the root directory.
- The active path to the working directory. Initially, the working directory will be located at the root of the file system.
- The constructor of the `Shell` class must properly initialize all the necessary information.

In the `Shell` class, which includes the representation of the active path, add the previous methods, with the same names and parameters as specified in the outline of this practice (adapted to the syntax of the chosen programming language). In addition to these, you can add any attributes and methods you consider necessary to solve this problem.
**NOTE:** The name of the class in both C++ and Java must be as proposed (Shell). Furthermore, in the case of C++, its interface will be defined in a file named 'shell.h', which is the only one included from the main program.

**Recommendations**
The difficulty of this practice depends on reasonably simple decisions made initially. Here are some tips to complete the practice without getting stuck on issues unrelated to the subject:
- To represent a directory, you will likely need a collection or vector of elements. Do not use basic arrays in C++ or Java because you need that collection to grow dynamically. We recommend using some of the standard library containers of each language. Specifically, consider using lists, vectors, or dictionaries. To store the elements of a directory, you should use a single collection that holds all the elements contained in the directory.
- Since the creation and deletion of files, directories, and links in this practice require the allocation and deletion of dynamic memory, automatic management of dynamic memory greatly facilitates the implementation. The Java language offers this management directly through automatic garbage collection. In C++, the deletion of this memory must be done manually, so for this practice, we strongly recommend using smart pointers from the C++ library (particularly `std::shared_ptr`).
- Every class library should be thoroughly tested to ensure it meets the required functionality, including exceptional situations. As assistance, we provide a file (`Main.java` for Java or `main.cc` for C++) that contains code to help you test your application. Additionally, we recommend creating your own files to thoroughly test your data structure and the corresponding methods.

**Important Note:**
Your class library must compile (without any errors or warnings) with the main program (`main.cc` or `Main.java`) provided, without the need to modify that file. If it does not, the evaluation of this practice will result in a score of 0. This requires the names of the classes and corresponding methods to match what is requested in this outline.